share setup; verify test dir
