+++ test/server/shared/utils/stub-component.js
@@ -1,5 +1,3 @@
-'use strict';
-
 import React, { Component } from 'react';
 
 export default class Stub extends Component {
